[11.0] [FIX] account_analytic_default_account: do not associate acount_id with _onchange_product_id method
If we change the asset, it will change the account_id. Then if we associate in this module the execution of _onchange_product_id by dependency of account_id, the same method of the account_asset module will also be executed, so the selected asset will be changed by the one of the product that can be a different one.

I propose to change the account_id dependency to the _onchange_account_id method.